refactor: separate out hash fn
